src/requirements.txt
pyusb==1.0.2
Flask==0.12.2
